+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnodegl`.
 - Optional `backend_config` field to the `ngl_config` structure (ABI break)
 - External OpenGL context support
 - `ngl-diff` tool to visually compare videos
+- `UniformColor.space` colorspace configuration (sRGB, HSL, HSV)
 
 ### Fixed
 - Various memory leaks